Rename chrome prefixed files as aren't browser specific
