+++ .gitignore
@@ -11,7 +11,6 @@ public/css/*.*
 *.swo
 build/firefox
 build/chrome
-build/safari
 build/test
 test/background.js
 shared/js/content-scope/sjcl.js
