+++ shared/js/background/classes/tab.es6.js
@@ -30,7 +30,7 @@ const Site = require('./site.es6')
 const Tracker = require('./tracker.es6')
 const HttpsRedirects = require('./https-redirects.es6')
 const Companies = require('../companies.es6')
-const browserWrapper = require('./../$BROWSER-wrapper.es6')
+const browserWrapper = require('./../wrapper.es6')
 const webResourceKeyRegex = /.*\?key=(.*)/
 
 class Tab {
