+++ shared/js/background/companies.es6.js
@@ -1,6 +1,6 @@
 const TopBlocked = require('./classes/top-blocked.es6')
 const Company = require('./classes/company.es6')
-const browserWrapper = require('./$BROWSER-wrapper.es6')
+const browserWrapper = require('./wrapper.es6')
 const migrate = require('./migrate.es6')
 
 const Companies = (() => {
