+++ shared/js/background/events.es6.js
@@ -304,7 +304,7 @@ chrome.omnibox.onInputEntered.addListener(function (text) {
 /**
  * MESSAGES
  */
-const browserWrapper = require('./chrome-wrapper.es6')
+const browserWrapper = require('./wrapper.es6')
 const {
     REFETCH_ALIAS_ALARM,
     fetchAlias,
