+++ shared/js/background/pixel.es6.js
@@ -5,7 +5,7 @@
  *
  */
 const load = require('./load.es6')
-const browserWrapper = require('./$BROWSER-wrapper.es6')
+const browserWrapper = require('./wrapper.es6')
 const settings = require('./settings.es6')
 const parseUserAgentString = require('../shared-utils/parse-user-agent-string.es6')
