+++ shared/js/background/storage/tds.es6.js
@@ -2,7 +2,7 @@ const load = require('./../load.es6')
 const Dexie = require('dexie')
 const constants = require('../../../data/constants')
 const settings = require('./../settings.es6')
-const browserWrapper = require('./../$BROWSER-wrapper.es6')
+const browserWrapper = require('./../wrapper.es6')
 
 class TDSStorage {
     constructor () {
