+++ shared/js/background/wrapper.es6.js
