+++ shared/js/ui/models/background-message.es6.js
@@ -1,5 +1,5 @@
 const Parent = window.DDG.base.Model
-const browserUIWrapper = require('./../base/$BROWSER-ui-wrapper.es6.js')
+const browserUIWrapper = require('./../base/ui-wrapper.es6.js')
 
 /**
  * Background messaging is done via two methods:
