+++ shared/js/ui/models/search.es6.js
@@ -1,5 +1,5 @@
 const Parent = window.DDG.base.Model
-const browserUIWrapper = require('./../base/$BROWSER-ui-wrapper.es6.js')
+const browserUIWrapper = require('./../base/ui-wrapper.es6.js')
 
 function Search (attrs) {
     Parent.call(this, attrs)
