+++ shared/js/ui/pages/mixins/index.es6.js
@@ -1,4 +1,4 @@
 module.exports = {
-    setBrowserClassOnBodyTag: require('./$BROWSER-set-browser-class.es6.js'),
+    setBrowserClassOnBodyTag: require('./set-browser-class.es6.js'),
     parseQueryString: require('./parse-query-string.es6.js')
 }
