+++ shared/js/ui/pages/mixins/set-browser-class.es6.js
