+++ unit-test/background/classes/site.es6.js
@@ -1,5 +1,5 @@
 const Site = require('../../../shared/js/background/classes/site.es6')
-const browserWrapper = require('../../../shared/js/background/chrome-wrapper.es6')
+const browserWrapper = require('../../../shared/js/background/wrapper.es6')
 const load = require('./../../helpers/utils.es6')
 const fakeBrokenSites = require('./../../data/brokensites')
 const tdsStorage = require('../../../shared/js/background/storage/tds.es6')
