+++ unit-test/background/click-to-load-tests.js
@@ -1,4 +1,4 @@
-const chromeWrapper = require('../../shared/js/background/chrome-wrapper.es6.js')
+const chromeWrapper = require('../../shared/js/background/wrapper.es6.js')
 const tds = require('../../shared/js/background/trackers.es6')
 const tdsStorage = require('../../shared/js/background/storage/tds.es6')
 const tdsStorageStub = require('./../helpers/tds.es6')
